// source/frontend/src/models/api/generated/ApiGen_CodeTypes_LeaseInsuranceTypes.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
export enum ApiGen_CodeTypes_LeaseInsuranceTypes {
  AIRCRAFT = 'AIRCRAFT',
  GENERAL = 'GENERAL',
  MARINE = 'MARINE',
  ACCIDENT = 'ACCIDENT',
  UAVDRONE = 'UAVDRONE',
  VEHICLE = 'VEHICLE',
  OTHER = 'OTHER',
}